// src/app/model/Student.ts
export interface Student{
  id: number;
  fname: String;
  lname: String;
  oname: String;
  line1: String;
  line2: String;
  contact: String;
  email: String;
  nic: String;
  date: number;
  month: number;
  year: number;
  gender: String;
  dsDivision: String;
  gnDivision: String;
  language: String;
  experience: String;
  isFollowed: boolean;
  courseDuration: String;
  hasImage: boolean;
  imagePath: String
}